Updating Cmake build a binary that doesn't segfault
The segfault issue from years ago didn't come from Mesa3D or Raylib, but it came from bad Cmake instructions, which compiled the project, but caused a segfault.
You can confirm this is the reason because if you build the project manually with gcc then it actually doesn't get this runtime error.